refactor:优化代码和项目结构，降低代码间耦合度，优化逻辑复杂度，提高代码运行效率
