+++ main.go
@@ -3,58 +3,64 @@ package main
 import (
 	"Bilibili_Downloader/internal"
 	"Bilibili_Downloader/internal/sso"
-	"Bilibili_Downloader/internal/toolkit"
-	"Bilibili_Downloader/internal/update"
 	"Bilibili_Downloader/internal/video_processing"
 	"Bilibili_Downloader/pkg/httpclient"
+	"Bilibili_Downloader/pkg/toolkit"
+	"Bilibili_Downloader/pkg/toolkit/data_struct"
 	"bufio"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"time"
 )
+
+func CatchVideoInfo(VideoInfoUrl string) (*data_struct.VideoInfoResponse, error) {
+	//获取视频信息
+	data, err := internal.CatchData(VideoInfoUrl)
+	if err != nil {
+		log.Printf("获取视频信息数据错误: %v\n\n", err)
+		fmt.Println("视频信息数据获取异常，请检查网络连接或前往log文件查看详情.")
+		return nil, err
+	}
+
+	//视频信息反序列化
+	Response, err := internal.ProcessResponse(data, 0)
+	if err != nil {
+		log.Printf("处理视频详情发生错误: %v\n\n", err)
+		fmt.Println("视频详情数据处理发生错误，请携带log文件向开发者反馈！")
+		return nil, err
+	}
+	videoInfoResponse, ok := Response.(*data_struct.VideoInfoResponse)
+	if !ok {
+		log.Println("视频详情数据类型断言失败")
+		fmt.Println("程序运行发生异常，请携带log日志文件联系开发者！")
+		return nil, fmt.Errorf("视频详情数据类型断言失败")
+	}
+	return videoInfoResponse, nil
+}
 
 func main() {
 	// 创建一个新的读取器
 	reader := bufio.NewReader(os.Stdin)
 
-	//初始化日志文件
-	logFile, err := os.OpenFile("debug.log", os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0666)
-	if err != nil {
-		log.Fatalf("无法打开日志文件: %v", err)
-	}
+	logFile := internal.InitLog()
 	defer func() {
 		if err := logFile.Close(); err != nil {
 			log.Println("log文件close失败:", err)
 		}
 	}()
-	// 将日志输出设置到文件
-	log.SetOutput(logFile)
-	// 设置日志前缀和格式
-	log.SetFlags(log.Ldate | log.Ltime | log.Lshortfile)
-	log.Println("程序运行，开始日志记录")
 
 	defer func() {
 		fmt.Printf("程序执行完毕，请按Enter键退出...")
-		// 读取一个字符
 		_, _ = reader.ReadString('\n')
 		log.Println("程序执行完毕，正常退出")
 	}()
 
 	//提交清理计划
-	defer func() {
-		if err := toolkit.RemoveCacheDir(); err != nil {
-			toolkit.ClearScreen()
-			log.Println("缓存目录清理失败:", err)
-			fmt.Println("缓存目录清理失败，确认需清理时可手动清理或重新运行程序")
-		} else {
-			fmt.Println("缓存目录清理完毕")
-		}
-	}()
+	defer internal.CacheClean()
 
 	//初始化客户端
-	if success := httpclient.Init(); !success {
+	if !httpclient.Init() {
 		err := sso.HandleQRCodeLogin()
 		if err != nil {
 			fmt.Println("处理二维码登录失败:", err)
@@ -62,23 +68,7 @@ func main() {
 		}
 	}
 
-	//检查更新
-	if check, newProgramName := update.CheckAndUpdate(); check == -1 {
-		fmt.Println("程序运行异常，请携带log日志联系开发者反馈")
-		return
-	} else if check == 1 {
-		// 创建一个临时程序来替换老版本
-		newProgram := fmt.Sprintf(".\\%s", newProgramName)
-		cmd := exec.Command("cmd.exe", "/K", "start", "", newProgram, "--update", os.Args[0])
-		if err := cmd.Start(); err != nil {
-			log.Println("启动新程序失败:", err)
-			fmt.Println("启动更新程序失败")
-			return
-		}
-		os.Exit(0)
-	} else {
-		log.Println("完成检查更新过程")
-	}
+	internal.HandleUpdate()
 
 	//主逻辑循环
 	for {
@@ -86,31 +76,13 @@ func main() {
 		BVid := toolkit.CatchAndCheckBVid()
 		VideoInfoUrl := fmt.Sprintf("https://api.bilibili.com/x/web-interface/view?bvid=%s", BVid)
 
-		//获取视频信息
-		data, err := internal.CatchData(VideoInfoUrl)
+		videoInfoResponse, err := CatchVideoInfo(VideoInfoUrl)
 		if err != nil {
-			log.Printf("获取视频信息数据错误: %v\n\n", err)
-			fmt.Println("视频信息数据获取异常，请检查网络连接或前往log文件查看详情.")
-			return
-		}
-
-		//视频信息反序列化
-		Response, err := internal.ProcessResponse(data, 0)
-		if err != nil {
-			log.Printf("处理视频详情发生错误: %v\n\n", err)
-			fmt.Println("视频详情数据处理发生错误，请携带log文件向开发者反馈！")
-			return
-		}
-		videoInfoResponse, ok := Response.(*toolkit.VideoInfoResponse)
-		if !ok {
-			log.Println("视频详情数据类型断言失败")
-			fmt.Println("程序运行发生异常，请携带log日志文件联系开发者！")
 			break
 		}
-
 		//打印视频详细信息并进行确认
 		toolkit.ClearScreen()
-		internal.ConfirmVideoExplanation(videoInfoResponse)
+		toolkit.ConfirmVideoExplanation(videoInfoResponse)
 
 		Default := -2
 		var actions map[string]map[string]int64
@@ -121,10 +93,10 @@ func main() {
 			if toolkit.YesOrNo() {
 				Default = 0
 				if len(videoInfoResponse.Data.Pages) > 1 {
-					internal.PrintDiversityInformationPart2(videoInfoResponse)
+					toolkit.PrintDiversityInformationPart2(videoInfoResponse)
 					part = 2
 				} else {
-					internal.PrintDiversityInformationPart1(videoInfoResponse)
+					toolkit.PrintDiversityInformationPart1(videoInfoResponse)
 					part = 1
 				}
 				if actions = toolkit.GetMaps(videoInfoResponse, part); actions == nil {
@@ -150,7 +122,7 @@ func main() {
 			for bvid, cid := range video {
 				//请求视频取流地址
 				DownloadURL := fmt.Sprintf("https://api.bilibili.com/x/player/wbi/playurl?bvid=%s&cid=%d&fnval=4048", bvid, cid)
-				data, err = internal.CatchData(DownloadURL)
+				data, err := internal.CatchData(DownloadURL)
 				if err != nil {
 					log.Printf("获取下载信息数据发生错误: %v\n\n", err)
 					fmt.Println("视频下载信息获取异常，请检查网络连接或前往log文件查看详情.")
@@ -164,7 +136,7 @@ func main() {
 					fmt.Println("视频下载信息处理发生错误，请携带log文件向开发者反馈！")
 					return
 				}
-				downloadInfoResponse, ok := newResponse.(*toolkit.DownloadInfoResponse)
+				downloadInfoResponse, ok := newResponse.(*data_struct.DownloadInfoResponse)
 				if !ok {
 					log.Println("视频下载数据类型断言失败")
 					fmt.Println("程序运行发生异常，请携带log日志文件联系开发者！")
