+++ pkg/toolkit/data_struct/download_info_struct.go
@@ -1,4 +1,4 @@
-package toolkit
+package data_struct
 
 type SegmentBase struct {
 	Initialization string `json:"Initialization"`
